feat: extract activity types into src/components/activity/types.ts
Move LogEntry, FileChange, ArtifactEntry, ManifestPayload, ArtifactManifest,
and IterationUsage interfaces from AppView.tsx into a dedicated types module.
Re-export them from AppView.tsx to preserve existing public API.

Aloop-Agent: build
Aloop-Iteration: 2
Aloop-Session: orchestrator-20260321-172932-issue-155-20260323-133710
